+++ front-end/src/components/Content/Content.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import * as styles from './Content.scss';
 
 import List from '../List';
-import Thumbnails from '../Thumbnails';
+import MostPopularBooks from '../MostPopularBooks';
 
 import Book from '../../models/Book';
 import Reader from '../../models/Reader';
@@ -16,40 +16,47 @@ import BookListItem from '../List/ListItem/BookListItem';
 
 import img from '../../assets/images/avatar.png';
 
-const Content: React.FC = () => (
-  <>
-    <main className={styles.main}>
-      <Thumbnails />
-    </main>
-    <aside className={styles.aside}>
-      <List<Reader>
-        title="Top czytelnicy"
-        items={readers}
-        render={(reader: Reader): JSX.Element => (
-          <ReaderListItem
-            key={reader.id}
-            id={reader.id}
-            firstname={reader.name}
-            lastname={reader.lastname}
-            avatarimg={img}
-          />
-        )}
-      />
-      <List<Book>
-        title="Najnowsze książki"
-        items={books}
-        render={(book: Book): JSX.Element => (
-          <BookListItem
-            key={book.id}
-            id={book.id}
-            title={book.title}
-            author={book.author}
-            coverimg="https://www.swiatksiazki.pl/media/catalog/product/cache/a946e6dbdb55333e1c3d566a3e38b923/7/2/7299906600872.jpg"
-          />
-        )}
-      />
-    </aside>
-  </>
-);
+const Content: React.FC = () => {
+  const slicedBooks = books.slice(books.length - 8, books.length);
+
+  return (
+    <>
+      <main className={styles.main}>
+        <div className={styles.mainWrapper}>
+          <h3 className={styles.mainHeader}>Najpopularniejsze ksiazki</h3>
+          <MostPopularBooks slicedBooks={slicedBooks} />
+        </div>
+      </main>
+      <aside className={styles.aside}>
+        <List<Reader>
+          title="Top czytelnicy"
+          items={readers}
+          render={(reader: Reader): JSX.Element => (
+            <ReaderListItem
+              key={reader.id}
+              id={reader.id}
+              firstname={reader.name}
+              lastname={reader.lastname}
+              avatarimg={img}
+            />
+          )}
+        />
+        <List<Book>
+          title="Najnowsze książki"
+          items={books}
+          render={(book: Book): JSX.Element => (
+            <BookListItem
+              key={book.id}
+              id={book.id}
+              title={book.title}
+              author={book.author}
+              coverimg="https://www.swiatksiazki.pl/media/catalog/product/cache/a946e6dbdb55333e1c3d566a3e38b923/7/2/7299906600872.jpg"
+            />
+          )}
+        />
+      </aside>
+    </>
+  );
+};
 
 export default Content;
