+++ front-end/src/components/MostPopularBooks/BookCard/index.ts
@@ -1,0 +1,2 @@
+import BookCard from './BookCard';
+export default BookCard;
