+++ front-end/src/components/MostPopularBooks/index.ts
@@ -1,0 +1,2 @@
+import MostPopularBooks from './MostPopularBooks';
+export default MostPopularBooks;
