--- front-end/src/components/Thumbnails/Thumbnail/index.ts
@@ -1,2 +1,0 @@
-import Thumbnail from './Thumbnail';
-export default Thumbnail;
