--- front-end/src/components/Thumbnails/index.ts
@@ -1,2 +1,0 @@
-import Thumbnails from './Thumbnails';
-export default Thumbnails;
